[260115] refactor : buyer/seller 삭제 작업 진행중
